Add vite + inject commit
