+++ npm/eslint.config.mjs
@@ -19,6 +19,6 @@ export default tseslint.config(
     },
   },
   {
-    ignores: ['dist', 'eslint.config.mjs'],
+    ignores: ['dist', 'eslint.config.mjs', 'vite.config.ts'],
   },
 );
