+++ npm/vite-env.d.ts
@@ -1,0 +1,4 @@
+/// <reference types="vite/client" />
+declare const __COMMIT_HASH__: string;
+declare const __COMMIT_DATE__: string;
+declare const __GIT_ORIGIN_URL__: string;
